Create .cpp file
